stable-diffusion-3.5-text2image-lora/requirements.txt
# Core PyTorch and ML libraries
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Hugging Face ecosystem
accelerate>=0.25.0
transformers>=4.35.0
diffusers>=0.25.0
peft>=0.7.0
datasets>=2.15.0

# Image processing and utilities
Pillow>=9.0.0
tqdm>=4.64.0

# Optional dependencies for advanced features
# Uncomment as needed:

# For experiment tracking
# wandb>=0.16.0
# tensorboard>=2.14.0

# For memory-efficient training
# bitsandbytes>=0.41.0

# For advanced optimizers
# prodigyopt>=1.0

# For development (uncomment if needed)
# black>=23.0.0
# flake8>=6.0.0
# pytest>=7.0.0
# pre-commit>=3.0.0